Monitor is done :+1: . Advanced JS externs files.
- Added separated externs files. To be parsed from GitHub.
- Fixed HTML5 syntax  for form buttons with labels.
- TODO: Minify JS.

+++ inc/classes/accountactivation.class.php
@@ -406,7 +406,7 @@ class AccountActivation extends Panes {
 		}
 
 		//* In-house transient cache.
-		$timestamp = ceil( time() / $divider );
+		$timestamp = (int) ceil( time() / $divider );
 
 		//* Return cached status within 2 hours.
 		if ( ! $doing_activation && $timestamp === $status['timestamp'] )
